backend/pyproject.toml
[project]
name = "something"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.116.1",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
]
